backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
sqlalchemy
asyncpg
psycopg2-binary
python-multipart
python-dotenv
selenium
openrouteservice
webdriver-manager
alembic
httpx
